Updating BoostParts to 1.42 and now using bind and lambda, too

+++ src/BoostParts/boost/throw_exception.hpp
@@ -22,6 +22,7 @@
 
 #include <boost/exception/detail/attribute_noreturn.hpp>
 #include <boost/detail/workaround.hpp>
+#include <boost/config.hpp>
 #include <exception>
 
 #if !defined( BOOST_EXCEPTION_DISABLE ) && defined( __BORLANDC__ ) && BOOST_WORKAROUND( __BORLANDC__, BOOST_TESTED_AT(0x593) )
@@ -38,7 +39,7 @@
 # define BOOST_THROW_EXCEPTION(x) ::boost::throw_exception(::boost::enable_error_info(x) <<\
     ::boost::throw_function(BOOST_CURRENT_FUNCTION) <<\
     ::boost::throw_file(__FILE__) <<\
-    ::boost::throw_line((int)__LINE__))
+    ::boost::throw_line(__LINE__))
 #else
 # define BOOST_THROW_EXCEPTION(x) ::boost::throw_exception(x)
 #endif
